+++ demisto_sdk/commands/validate/validate_manager.py
@@ -329,7 +329,8 @@ class ValidateManager:
         elif file_type == FileType.BETA_INTEGRATION:
             return self.validate_beta_integration(structure_validator, pack_error_ignore_list)
 
-        elif file_type == FileType.IMAGE:
+        # Validate only images of packs
+        elif file_type == FileType.IMAGE and file_path.endswith('_image.png'):
             return self.validate_image(file_path, pack_error_ignore_list)
 
         # incident fields and indicator fields are using the same scheme.
@@ -841,8 +842,9 @@ class ValidateManager:
                 file_path = file_data[2]
 
             # if the file is a code file - change path to the associated yml path to trigger release notes validation.
-            if find_type(file_path) in [FileType.POWERSHELL_FILE, FileType.PYTHON_FILE] and file_status.lower() != 'd' \
-                    and not (file_path.endswith('_test.py') or file_path.endswith('.Tests.ps1')):
+            if file_status.lower() != 'd' and \
+                find_type(file_path) in [FileType.POWERSHELL_FILE, FileType.PYTHON_FILE] and \
+                    not (file_path.endswith('_test.py') or file_path.endswith('.Tests.ps1')):
                 # naming convention - code file and yml file in packages must have same name.
                 file_path = os.path.splitext(file_path)[0] + '.yml'
 
@@ -852,7 +854,7 @@ class ValidateManager:
                 continue
 
             # identify deleted files
-            if file_status.lower() == 'd' and find_type(file_path) and not file_path.startswith('.'):
+            if file_status.lower() == 'd' and not file_path.startswith('.'):
                 deleted_files.add(file_path)
 
             # ignore directories